Create README.md
Demo app to demonstrate drag-and-drop issues.

Issue 1: When the display scale is other than 100%, then the cursor vs. the drop target is calculated wrong.
Steps 1: Set display scaling to 100% and drag-and-drop the (orange) source in the green target -> works fine.
Steps 2: Set display scaling to other than 100% and drag-and-drop the (orange) source in the green target -> depending in scaling, window position and size, the target area is somewhere other that the green box (unpredicatbly where).

Issue 2: Drag-and-drop will fataly (unrecoverably) crash the app after 1 or more drag-and-drop operations.
Step 1: Set display scaling to 100%
Step 2: Repeat drag-and-drop until the application crashes, which might be right the first time, or after many (over 50) times. But eventualy, it will crash.